remove old.rs and put contents of new.rs to circuit/mod.rs
